<Fran> Added Main Menu
