--- Assets/Menu.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 33945e759ec6a594db5d2eda466ffbc3
