--- Assets/MusicIntroScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 432045dd372a75c49ae5f24bc23bd3c7
