--- Assets/RoomManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7387d4ee7340def46894266ff274956f
