+++ Assets/Scripts/Menu.cs
@@ -1,0 +1,23 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class Menu : MonoBehaviour
+{
+    // Start is called once before the first execution of Update after the MonoBehaviour is created
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+    public void LoadIntro()
+    {
+        Debug.Log("entra");
+        SceneManager.LoadScene(1);
+    }
+}
