+++ Assets/Scripts/MenuManager.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class MenuManager : MonoBehaviour
+{
+    public void StartGame()
+    {
+        SceneManager.LoadScene("Intro Timeline");
+    }
+}
