+++ Assets/Scripts/funcionarScript.cs
@@ -1,0 +1,22 @@
+using UnityEngine;
+
+public class funcionarScript : MonoBehaviour
+{
+    // Start is called once before the first execution of Update after the MonoBehaviour is created
+    void Start()
+    {
+
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        
+    }
+
+    public void funciona()
+    {
+        Debug.Log("Funciona!");
+    }
+}
+
