+++ Assets/Scripts/funcionarScript.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: abcfff8fe89525547b3ecc5cf8330c5a
