--- Assets/Signal.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 092b9869a3f16c240876e15d8506c50d
