--- Assets/backToMenu.cs
@@ -1,22 +1,0 @@
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class backToMenu : MonoBehaviour
-{
-    // Start is called once before the first execution of Update after the MonoBehaviour is created
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
-    public void goBackToMenu()
-    {
-        SceneManager.LoadScene(0);
-    }
-}
