--- Assets/backToMenu.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7f17fce9d151af940ba3694899de534b
