--- Assets/canvasScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2bc38f9c7100fd64f9ffccd99536d703
