--- Assets/endingScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 5c00a4206757b5e43ab28fbf11c18cc8
