--- Assets/formScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a803d0888edc8e44e8a52b04d2c8d7eb
