--- Assets/ima.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 29bb2b721c85f2146ae0b0f306fa131e
